Добавляет класс прородителя Component
