+++ src/components/task.js
@@ -1,9 +1,10 @@
-import {getTaskTemplate} from '../src/get-template';
-import {createElement} from '../src/utils';
+import {getTaskTemplate} from '../templates/task-templates';
+import {Component} from './component';
 
 
-export default class Task {
+export default class Task extends Component {
   constructor(data) {
+    super();
     this._title = data.title;
     this._dueDate = data.dueDate;
     this._tags = data.tags;
@@ -11,25 +12,15 @@ export default class Task {
     this._color = data.color;
     this._repeatingDays = data.repeatingDays;
 
-    this._state = {
-      isDone: data.isDone,
-      isFavorite: data.isFavorite
-    };
-
-    this._element = null;
     this._onEdit = null;
   }
 
-  __onEditButtonClick() {
+  _onEditButtonClick() {
     return typeof this._onEdit === `function` && this._onEdit();
   }
 
   set onEdit(fn) {
     this._onEdit = fn;
-  }
-
-  get element() {
-    return this._element;
   }
 
   get template() {
@@ -38,23 +29,11 @@ export default class Task {
 
   bind() {
     this._element.querySelector(`.card__btn--edit`)
-      .addEventListener(`click`, this.__onEditButtonClick.bind(this));
+      .addEventListener(`click`, this._onEditButtonClick.bind(this));
   }
 
   unbind() {
     this._element.querySelector(`.card__btn--edit`)
-      .removeEventListener(`click`, this.__onEditButtonClick);
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
+      .removeEventListener(`click`, this._onEditButtonClick);
   }
 }
